requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
curl-cffi==0.13.0
dateparser==1.2.2
frozendict==2.4.7
frozenlist==1.8.0
gitdb==4.0.12
gitpython==3.1.46
idna==3.11
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
multidict==6.7.1
multitasking==0.0.12
narwhals==2.16.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
peewee==3.19.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.5.2
propcache==0.4.1
protobuf==6.33.5
pyarrow==21.0.0
pycparser==2.23
pycryptodome==3.23.0
pydeck==0.9.1
python-binance==1.0.34
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
yarl==1.22.0
yfinance==1.1.0
